requirements.txt
Django
djangorestframework
markdown
django-filter
firebase-admin
django-storages
boto3
blurhash-python
psycopg2-binary